Add decompose module
